Module Auto

This files implements auto and related automation tactics

val compute_secvars : Proofview.Goal.t -> Names.Id.Pred.t

val default_search_depth : int Stdlib.ref

val auto_flags_of_state : TransparentState.t -> Unification.unify_flags

val connect_hint_clenv : Hints.hint -> Proofview.Goal.t -> Clenv.clausenv * EConstr.constr

val unify_resolve : Unification.unify_flags -> Hints.hint -> unit Proofview.tactic

Try unification with the precompiled clause, then use registered Apply

val conclPattern : EConstr.constr -> Pattern.constr_pattern option -> Genarg.glob_generic_argument -> unit Proofview.tactic

val auto : ?⁠debug:Hints.debug -> int -> Tactypes.delayed_open_constr list -> Hints.hint_db_name list -> unit Proofview.tactic

val new_auto : ?⁠debug:Hints.debug -> int -> Tactypes.delayed_open_constr list -> Hints.hint_db_name list -> unit Proofview.tactic

val default_auto : unit Proofview.tactic

auto with default search depth and with the hint database "core"

val full_auto : ?⁠debug:Hints.debug -> int -> Tactypes.delayed_open_constr list -> unit Proofview.tactic

auto with all hint databases

val new_full_auto : ?⁠debug:Hints.debug -> int -> Tactypes.delayed_open_constr list -> unit Proofview.tactic

auto with all hint databases and doing delta

val default_full_auto : unit Proofview.tactic

auto with default search depth and with all hint databases

val gen_auto : ?⁠debug:Hints.debug -> int option -> Tactypes.delayed_open_constr list -> Hints.hint_db_name list option -> unit Proofview.tactic

The generic form of auto (second arg None means all bases)

val h_auto : ?⁠debug:Hints.debug -> int option -> Tactypes.delayed_open_constr list -> Hints.hint_db_name list option -> unit Proofview.tactic

The hidden version of auto

val trivial : ?⁠debug:Hints.debug -> Tactypes.delayed_open_constr list -> Hints.hint_db_name list -> unit Proofview.tactic

val gen_trivial : ?⁠debug:Hints.debug -> Tactypes.delayed_open_constr list -> Hints.hint_db_name list option -> unit Proofview.tactic

val full_trivial : ?⁠debug:Hints.debug -> Tactypes.delayed_open_constr list -> unit Proofview.tactic

val h_trivial : ?⁠debug:Hints.debug -> Tactypes.delayed_open_constr list -> Hints.hint_db_name list option -> unit Proofview.tactic