Module ComAssumption

Parameters/Assumptions

val do_assumptions : program_mode:bool -> poly:bool -> scope:DeclareDef.locality -> kind:Decls.assumption_object_kind -> Declaremods.inline -> (Constrexpr.ident_decl list * Constrexpr.constr_expr) Vernacexpr.with_coercion list -> unit

val declare_assumption : Vernacexpr.coercion_flag -> poly:bool -> scope:DeclareDef.locality -> kind:Decls.assumption_object_kind -> Constr.types -> Entries.universes_entry -> UnivNames.universe_binders -> Impargs.manual_implicits -> Glob_term.binding_kind -> Declaremods.inline -> Names.variable CAst.t -> Names.GlobRef.t * Univ.Instance.t

returns false if the assumption is neither local to a section, nor in a module type and meant to be instantiated.

val context : poly:bool -> Constrexpr.local_binder_expr list -> unit

returns false if, for lack of section, it declares an assumption (unless in a module type).

val do_primitive : Names.lident -> CPrimitives.op_or_type -> Constrexpr.constr_expr option -> unit