Module Conv_oracle

type oracle

val empty : oracle

val oracle_order : ('a -> Names.Constant.t) -> oracle -> bool -> 'a Names.tableKey -> 'a Names.tableKey -> bool

Order on section paths for unfolding. If oracle_order kn1 kn2 is true, then unfold kn1 first. Note: the oracle does not introduce incompleteness, it only tries to postpone unfolding of "opaque" constants.

type level =

| Expand

| Level of int

| Opaque

Priority for the expansion of constant in the conversion test. * Higher levels means that the expansion is less prioritary. * (And Expand stands for -oo, and Opaque +oo.) * The default value (transparent constants) is Level 0.

val transparent : level

val is_transparent : level -> bool

Check whether a level is transparent

val get_strategy : oracle -> Names.Constant.t Names.tableKey -> level

val set_strategy : oracle -> Names.Constant.t Names.tableKey -> level -> oracle

Sets the level of a constant. * Level of RelKey constant cannot be set.

val fold_strategy : (Names.Constant.t Names.tableKey -> level -> 'a -> 'a) -> oracle -> 'a -> 'a

Fold over the non-transparent levels of the oracle. Order unspecified.

val get_transp_state : oracle -> TransparentState.t