Module Coqcargs

type compilation_mode =

| BuildVo

| BuildVio

| Vio2Vo

| BuildVos

| BuildVok

Compilation modes:

BuildVo : process statements and proofs (standard compilation), and also output an empty .vos file and .vok file

BuildVio : process statements, delay proofs in futures

Vio2Vo : load delayed proofs and process them

BuildVos : process statements, and discard proofs, and load .vos files for required libraries

BuildVok : like BuildVo, but load .vos files for required libraries

When loading the .vos version of a required library, if the file exists but is empty, then we attempt to load the .vo version of that library. This trick is useful to avoid the need for the user to compile .vos version when an up to date .vo version is already available.

type t = {

compilation_mode : compilation_mode;

compile_list : (string * bool) list;

compilation_output_name : string option;

vio_checking : bool;

vio_tasks : (int list * string) list;

vio_files : string list;

vio_files_j : int;

echo : bool;

outputstate : string option;

glob_out : Dumpglob.glob_output;

}

val default : t

val parse : string list -> t