Module Declare.Obls

type 'a obligation_body =

| DefinedObl of 'a

| TermObl of Constr.constr

module Obligation : sig ... end

type obligations = {

obls : Obligation.t array;

remaining : int;

}

type fixpoint_kind =

| IsFixpoint of Names.lident option list

| IsCoFixpoint

module ProgramDecl : sig ... end

val declare_obligation : ProgramDecl.t -> Obligation.t -> uctx:UState.t -> types:Constr.types option -> body:Constr.types -> bool * Obligation.t

declare_obligation prg obl ~uctx ~types ~body Save an obligation obl for program definition prg

module State : sig ... end

val declare_definition : ProgramDecl.t -> Names.GlobRef.t

type progress =

| Remain of int
n obligations remaining

| Dependent
Dependent on other definitions

| Defined of Names.GlobRef.t
Defined as id

Resolution status of a program

type obligation_resolver = Names.Id.t option -> Int.Set.t -> unit Proofview.tactic option -> progress

type obligation_qed_info = {

name : Names.Id.t;

num : int;

auto : obligation_resolver;

}

val update_obls : ProgramDecl.t -> Obligation.t array -> int -> progress

update_obls prg obls n progress What does this do?

val check_solved_obligations : what_for:Pp.t -> unit

Check obligations are properly solved before closing the what_for section / module

val obl_substitution : bool -> Obligation.t array -> Int.Set.t -> (Names.Id.t * (Constr.types * Constr.types)) list

val dependencies : Obligation.t array -> int -> Int.Set.t

val stm_get_fix_exn : (unit -> Exninfo.iexn -> Exninfo.iexn) Stdlib.ref