Module DeclareDef

type locality =

| Discharge

| Global of Declare.import_status

module Hook : sig ... end

Declaration hooks

val declare_entry : name:Names.Id.t -> scope:locality -> kind:Decls.logical_kind -> ?⁠hook:Hook.t -> ?⁠obls:(Names.Id.t * Constr.t) list -> impargs:Impargs.manual_implicits -> uctx:UState.t -> Evd.side_effects Declare.proof_entry -> Names.GlobRef.t

Declare an interactively-defined constant

val declare_definition : name:Names.Id.t -> scope:locality -> kind:Decls.logical_kind -> opaque:bool -> impargs:Impargs.manual_implicits -> udecl:UState.universe_decl -> ?⁠hook:Hook.t -> ?⁠obls:(Names.Id.t * Constr.t) list -> poly:bool -> ?⁠inline:bool -> types:EConstr.t option -> body:EConstr.t -> ?⁠fix_exn:(Exninfo.iexn -> Exninfo.iexn) -> Evd.evar_map -> Names.GlobRef.t

Declares a non-interactive constant; body and types will be normalized w.r.t. the passed evar_map sigma. Universes should be handled properly, including minimization and restriction. Note that sigma is checked for unresolved evars, thus you should be careful not to submit open terms or evar maps with stale, unresolved existentials

val declare_assumption : ?⁠fix_exn:(Exninfo.iexn -> Exninfo.iexn) -> name:Names.Id.t -> scope:locality -> hook:Hook.t option -> impargs:Impargs.manual_implicits -> uctx:UState.t -> Entries.parameter_entry -> Names.GlobRef.t

module Recthm : sig ... end

val declare_mutually_recursive : opaque:bool -> scope:locality -> kind:Decls.logical_kind -> poly:bool -> uctx:UState.t -> udecl:UState.universe_decl -> ntns:Vernacexpr.decl_notation list -> rec_declaration:Constr.rec_declaration -> possible_indexes:int list list option -> ?⁠restrict_ucontext:bool -> Recthm.t list -> Names.GlobRef.t list

val prepare_obligation : ?⁠opaque:bool -> ?⁠inline:bool -> name:Names.Id.t -> poly:bool -> udecl:UState.universe_decl -> types:EConstr.t option -> body:EConstr.t -> Evd.evar_map -> Constr.constr * Constr.types * UState.t * RetrieveObl.obligation_info

val prepare_parameter : poly:bool -> udecl:UState.universe_decl -> types:EConstr.types -> Evd.evar_map -> Evd.evar_map * Entries.parameter_entry