Module DeclareDef.Hook

Declaration hooks

type t

module S : sig ... end

Hooks allow users of the API to perform arbitrary actions at proof/definition saving time. For example, to register a constant as a Coercion, perform some cleanup, update the search database, etc...

val make : (S.t -> unit) -> t

val call : ?⁠hook:t -> ?⁠fix_exn:Future.fix_exn -> S.t -> unit