Module Evd.FutureGoals

type t = private {

comb : Evar.t list;

principal : Evar.t option;
if Some e, e must be contained in future_comb. The evar e will inherit properties (now: the name) of the evar which will be instantiated with a term containing e.

}

val map_filter : (Evar.t -> Evar.t option) -> t -> t

Applies a function on the future goals

val filter : (Evar.t -> bool) -> t -> t

Applies a filter on the future goals