Module type Grammar.S

Signature type of the functor Grammar.GMake. The types and functions are almost the same than in generic interface, but:

Grammars are not values. Functions holding a grammar as parameter do not have this parameter yet.

The type parsable is used in function parse instead of the char stream, avoiding the possible loss of tokens.

The type of tokens (expressions and patterns) can be any type (instead of (string * string)); the module parameter must specify a way to show them as (string * string)

type te

type 'c pattern

module Parsable : sig ... end

val tokens : string -> (string option * int) list

module Entry : sig ... end

module rec Symbol : sig ... end

and Rule : sig ... end

and Rules : sig ... end

module Production : sig ... end

module Unsafe : sig ... end

val safe_extend : warning:(string -> unit) option -> 'a Entry.t -> Gramext.position option -> (string option * Gramext.g_assoc option * 'a Production.t list) list -> unit

val safe_delete_rule : 'a Entry.t -> ('a, _, 'f, 'r) Rule.t -> unit