Module Notation_term

type notation_constr =

| NRef of Names.GlobRef.t * Glob_term.glob_level list option

| NVar of Names.Id.t

| NApp of notation_constr * notation_constr list

| NHole of Evar_kinds.t * Namegen.intro_pattern_naming_expr * Genarg.glob_generic_argument option

| NList of Names.Id.t * Names.Id.t * notation_constr * notation_constr * bool

| NLambda of Names.Name.t * notation_constr option * notation_constr

| NProd of Names.Name.t * notation_constr option * notation_constr

| NBinderList of Names.Id.t * Names.Id.t * notation_constr * notation_constr * bool

| NLetIn of Names.Name.t * notation_constr * notation_constr option * notation_constr

| NCases of Constr.case_style * notation_constr option * (notation_constr * (Names.Name.t * (Names.inductive * Names.Name.t list) option)) list * (Glob_term.cases_pattern list * notation_constr) list

| NLetTuple of Names.Name.t list * Names.Name.t * notation_constr option * notation_constr * notation_constr

| NIf of notation_constr * Names.Name.t * notation_constr option * notation_constr * notation_constr

| NRec of Glob_term.glob_fix_kind * Names.Id.t array * (Names.Name.t * notation_constr option * notation_constr) list array * notation_constr array * notation_constr array

| NSort of Glob_term.glob_sort

| NCast of notation_constr * notation_constr Glob_term.cast_type

| NInt of Uint63.t

| NFloat of Float64.t

| NArray of notation_constr array * notation_constr * notation_constr

type scope_name = string

type tmp_scope_name = scope_name

type subscopes = tmp_scope_name option * scope_name list

type extended_subscopes = Constrexpr.notation_entry_level * subscopes

type constr_as_binder_kind =

| AsIdent

| AsName

| AsNameOrPattern

| AsStrictPattern

type notation_binder_source =

| NtnParsedAsPattern of bool

| NtnParsedAsIdent

| NtnParsedAsName

| NtnBinderParsedAsConstr of constr_as_binder_kind

| NtnParsedAsBinder

type notation_var_instance_type =

| NtnTypeConstr

| NtnTypeBinder of notation_binder_source

| NtnTypeConstrList

| NtnTypeBinderList

type notation_var_internalization_type =

| NtnInternTypeAny

| NtnInternTypeOnlyBinder

Type of variables when interpreting a constr_expr as a notation_constr: in a recursive pattern x..y, both x and y carry the individual type of each element of the list x..y

type interpretation = (Names.Id.t * (extended_subscopes * notation_var_instance_type)) list * notation_constr

This characterizes to what a notation is interpreted to

type reversibility_status =

| APrioriReversible

| HasLtac

| NonInjective of Names.Id.t list

type notation_interp_env = {

ninterp_var_type : notation_var_internalization_type Names.Id.Map.t;

ninterp_rec_vars : Names.Id.t Names.Id.Map.t;

}