Module Prettyp

val assumptions_for_print : Names.Name.t list -> Termops.names_context

val print_closed_sections : bool Stdlib.ref

val print_context : Environ.env -> Evd.evar_map -> bool -> int option -> Lib.library_segment -> Pp.t

val print_library_entry : Environ.env -> Evd.evar_map -> bool -> (Libobject.object_name * Lib.node) -> Pp.t option

val print_full_context : Environ.env -> Evd.evar_map -> Pp.t

val print_full_context_typ : Environ.env -> Evd.evar_map -> Pp.t

val print_full_pure_context : Environ.env -> Evd.evar_map -> Pp.t

val print_sec_context : Environ.env -> Evd.evar_map -> Libnames.qualid -> Pp.t

val print_sec_context_typ : Environ.env -> Evd.evar_map -> Libnames.qualid -> Pp.t

val print_judgment : Environ.env -> Evd.evar_map -> EConstr.unsafe_judgment -> Pp.t

val print_safe_judgment : Environ.env -> Evd.evar_map -> Safe_typing.judgment -> Pp.t

val print_eval : Reductionops.reduction_function -> Environ.env -> Evd.evar_map -> Constrexpr.constr_expr -> EConstr.unsafe_judgment -> Pp.t

val print_name : Environ.env -> Evd.evar_map -> Libnames.qualid Constrexpr.or_by_notation -> UnivNames.univ_name_list option -> Pp.t

val print_opaque_name : Environ.env -> Evd.evar_map -> Libnames.qualid -> Pp.t

val print_about : Environ.env -> Evd.evar_map -> Libnames.qualid Constrexpr.or_by_notation -> UnivNames.univ_name_list option -> Pp.t

val print_impargs : Libnames.qualid Constrexpr.or_by_notation -> Pp.t

val print_graph : unit -> Pp.t

Pretty-printing functions for classes and coercions

val print_classes : unit -> Pp.t

val print_coercions : unit -> Pp.t

val print_path_between : Classops.cl_typ -> Classops.cl_typ -> Pp.t

val print_canonical_projections : Environ.env -> Evd.evar_map -> Pp.t

val print_typeclasses : unit -> Pp.t

Pretty-printing functions for type classes and instances

val print_instances : Names.GlobRef.t -> Pp.t

val print_all_instances : unit -> Pp.t

val inspect : Environ.env -> Evd.evar_map -> int -> Pp.t

Locate

type 'a locatable_info = {

locate : Libnames.qualid -> 'a option;
Locate the most precise object with the provided name if any.

locate_all : Libnames.qualid -> 'a list;
Locate all objects whose name is a suffix of the provided name

shortest_qualid : 'a -> Libnames.qualid;
Return the shortest name in the current context

name : 'a -> Pp.t;
Data as printed by the Locate command

print : 'a -> Pp.t;
Data as printed by the Print command

about : 'a -> Pp.t;
Data as printed by the About command

}

Generic data structure representing locatable objects.

val register_locatable : string -> 'a locatable_info -> unit

Define a new type of locatable objects that can be reached via the corresponding generic vernacular commands. The string should be a unique name describing the kind of objects considered and that is added as a grammar command prefix for vernacular commands Locate.

val print_located_qualid : Libnames.qualid -> Pp.t

val print_located_term : Libnames.qualid -> Pp.t

val print_located_module : Libnames.qualid -> Pp.t

val print_located_other : string -> Libnames.qualid -> Pp.t

type object_pr = {

print_inductive : Names.MutInd.t -> UnivNames.univ_name_list option -> Pp.t;

print_constant_with_infos : Names.Constant.t -> UnivNames.univ_name_list option -> Pp.t;

print_section_variable : Environ.env -> Evd.evar_map -> Names.variable -> Pp.t;

print_syntactic_def : Environ.env -> Names.KerName.t -> Pp.t;

print_module : bool -> Names.ModPath.t -> Pp.t;

print_modtype : Names.ModPath.t -> Pp.t;

print_named_decl : Environ.env -> Evd.evar_map -> Constr.named_declaration -> Pp.t;

print_library_entry : Environ.env -> Evd.evar_map -> bool -> (Libobject.object_name * Lib.node) -> Pp.t option;

print_context : Environ.env -> Evd.evar_map -> bool -> int option -> Lib.library_segment -> Pp.t;

print_typed_value_in_env : Environ.env -> Evd.evar_map -> (EConstr.constr * EConstr.types) -> Pp.t;

print_eval : Reductionops.reduction_function -> Environ.env -> Evd.evar_map -> Constrexpr.constr_expr -> EConstr.unsafe_judgment -> Pp.t;

}

val set_object_pr : object_pr -> unit

val default_object_pr : object_pr