Module Univ.Level

Universes.

module UGlobal : sig ... end

Qualified global universe level

type t

Type of universe levels. A universe level is essentially a unique name that will be associated to constraints later on. A level can be local to a definition or global.

val set : t

val prop : t

val sprop : t

The set and prop universe levels.

val is_small : t -> bool

Is the universe set or prop?

val is_sprop : t -> bool

val is_prop : t -> bool

val is_set : t -> bool

Is it specifically Prop or Set

val compare : t -> t -> int

Comparison function

val equal : t -> t -> bool

Equality function

val hash : t -> int

val make : UGlobal.t -> t

val pr : t -> Pp.t

Pretty-printing

val to_string : t -> string

Debug printing

val var : int -> t

val var_index : t -> int option

val name : t -> UGlobal.t option