Module Usage

Prints the version number on the standard output.

val version : unit -> unit

val machine_readable_version : unit -> unit

extra arguments or options to print when asking usage for a given executable.

type specific_usage = {

executable_name : string;

extra_args : string;

extra_options : string;

}

Prints the generic part and specific part of usage for a given executable.

val print_usage : Stdlib.out_channel -> specific_usage -> unit