Module Vernacinterp

val interp : ?⁠verbosely:bool -> st:Vernacstate.t -> Vernacexpr.vernac_control -> Vernacstate.t

The main interpretation function of vernacular expressions

val interp_qed_delayed_proof : proof:Declare.Proof.proof_object -> pinfo:Declare.Proof.Proof_info.t -> st:Vernacstate.t -> control:Vernacexpr.control_flag list -> Vernacexpr.proof_end CAst.t -> Vernacstate.t

Execute a Qed but with a proof_object which may contain a delayed proof and won't be forced

val with_fail : st:Vernacstate.t -> (unit -> 'a) -> unit

with_fail ~st f runs f () and expects it to fail, otherwise it fails.

val test_mode : bool Stdlib.ref

Flag set when the test-suite is called. Its only effect to display verbose information for Fail

val get_default_proof_mode : unit -> Pvernac.proof_mode

Default proof mode set by `start_proof`

val proof_mode_opt_name : string list