Module Ssrmatching_plugin.G_ssrmatching

val cpattern : Ssrmatching.cpattern Pcoq.Entry.t

CS cpattern: (f _), (X in t), (t in X in t), (t as X in t)

val wit_cpattern : Ssrmatching.cpattern Genarg.uniform_genarg_type

val lcpattern : Ssrmatching.cpattern Pcoq.Entry.t

OS cpattern: f _, (X in t), (t in X in t), (t as X in t)

val wit_lcpattern : Ssrmatching.cpattern Genarg.uniform_genarg_type

val rpattern : Ssrmatching.rpattern Pcoq.Entry.t

OS rpattern: f _, in t, X in t, in X in t, t in X in t, t as X in t

val wit_rpattern : Ssrmatching.rpattern Genarg.uniform_genarg_type